extract ArmorSummary to standalone class
