+++ src/main/java/com/brandon3055/draconicevolution/client/handler/HudHandler.java
@@ -19,7 +19,7 @@ import org.lwjgl.opengl.GL11;
 import com.brandon3055.draconicevolution.client.gui.GuiHudConfig;
 import com.brandon3055.draconicevolution.client.utils.GuiHelper;
 import com.brandon3055.draconicevolution.common.handler.ConfigHandler;
-import com.brandon3055.draconicevolution.common.items.armor.CustomArmorHandler.ArmorSummary;
+import com.brandon3055.draconicevolution.common.items.armor.ArmorSummary;
 import com.brandon3055.draconicevolution.common.utils.IHudDisplayBlock;
 import com.brandon3055.draconicevolution.common.utils.IHudDisplayItem;
